+++ src/main/java/net/geforcemods/securitycraft/screen/KeycardReaderScreen.java
@@ -53,6 +53,7 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 	private static final Component EQUALS = Component.literal("=");
 	private static final Component GREATER_THAN_EQUALS = Component.literal(">=");
 	private static final int MAX_SIGNATURE = 99999;
+	private final Component signatureText = Utils.localize("gui.securitycraft:keycard_reader.signature");
 	private final Component keycardLevelsText = Utils.localize("gui.securitycraft:keycard_reader.keycard_levels");
 	private final Component linkText = Utils.localize("gui.securitycraft:keycard_reader.link");
 	private final Component levelMismatchInfo = Utils.localize("gui.securitycraft:keycard_reader.level_mismatch");
@@ -65,12 +66,11 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 	private int previousSignature;
 	private int signature;
 	private boolean[] acceptedLevels;
-	private Component signatureText;
 	private int signatureTextLength;
 	private int signatureTextStartX;
 	private Button minusThree, minusTwo, minusOne, reset, plusOne, plusTwo, plusThree;
 	private TogglePictureButton[] toggleButtons = new TogglePictureButton[5];
-	private EditBox usesTextField, usableByTextField;
+	private EditBox signatureTextField, usesTextField, usableByTextField;
 	private TextHoverChecker usesHoverChecker;
 	private Button setUsesButton;
 	private Button linkButton;
@@ -137,6 +137,13 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 			}
 		}
 
+		signatureTextLength = font.width(signatureText);
+		signatureTextStartX = imageWidth / 2 - signatureTextLength + 5;
+		signatureTextField = addRenderableWidget(new EditBox(font, leftPos + 96, topPos + 21, 40, 12, Component.empty()));
+		signatureTextField.setValue(leftPaddedSignature());
+		signatureTextField.setFilter(s -> s.matches("\\d*"));
+		signatureTextField.setMaxLength(5);
+		signatureTextField.setResponder(this::changeSignature);
 		minusThree = addRenderableWidget(new Button(leftPos + 22, buttonY, 24, buttonHeight, Component.literal("---"), b -> changeSignature(signature - 100), Button.DEFAULT_NARRATION));
 		minusTwo = addRenderableWidget(new Button(leftPos + 48, buttonY, 18, buttonHeight, Component.literal("--"), b -> changeSignature(signature - 10), Button.DEFAULT_NARRATION));
 		minusOne = addRenderableWidget(new Button(leftPos + 68, buttonY, 12, buttonHeight, Component.literal("-"), b -> changeSignature(signature - 1), Button.DEFAULT_NARRATION));
@@ -217,7 +224,7 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 	@Override
 	protected void renderLabels(GuiGraphics guiGraphics, int mouseX, int mouseY) {
 		guiGraphics.drawString(font, title, imageWidth / 2 - font.width(title) / 2, 6, CommonColors.DARK_GRAY, false);
-		guiGraphics.drawString(font, signatureText, imageWidth / 2 - font.width(signatureText) / 2, 23, CommonColors.DARK_GRAY, false);
+		guiGraphics.drawString(font, signatureText, signatureTextStartX, 23, CommonColors.DARK_GRAY, false);
 		guiGraphics.drawString(font, keycardLevelsText, 170 - font.width(keycardLevelsText), 56, CommonColors.DARK_GRAY, false);
 
 		//numbers infront of keycard levels buttons
@@ -301,7 +308,7 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 
 	@Override
 	public boolean mouseScrolled(double mouseX, double mouseY, double scrollX, double scrollY) {
-		if (isOwner && mouseX >= leftPos + signatureTextStartX && mouseY >= topPos + 23 && mouseX <= leftPos + signatureTextStartX + signatureTextLength && mouseY <= topPos + 43)
+		if (isOwner && mouseX >= leftPos + signatureTextStartX && mouseY >= topPos + 23 && mouseX <= leftPos + signatureTextStartX + signatureTextLength + 42 && mouseY <= topPos + 43)
 			changeSignature(signature + (int) Math.signum(scrollY));
 
 		return super.mouseScrolled(mouseX, mouseY, scrollX, scrollY);
@@ -332,15 +339,21 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 	}
 
 	public void changeSignature(int newSignature) {
+		changeSignature(newSignature, false);
+	}
+
+	public void changeSignature(String newSignature) {
+		if (newSignature != null && !newSignature.isEmpty())
+			changeSignature(Integer.parseInt(newSignature), true);
+	}
+
+	public void changeSignature(int newSignature, boolean throughTextField) {
 		boolean enablePlusButtons;
 		boolean enableMinusButtons;
 
 		if (isOwner)
 			signature = Mth.clamp(newSignature, 0, MAX_SIGNATURE); //keep between 0 and the max allowed (disallow negative numbers)
 
-		signatureText = Component.translatable("gui.securitycraft:keycard_reader.signature", StringUtils.leftPad("" + signature, 5, "0"));
-		signatureTextLength = font.width(signatureText);
-		signatureTextStartX = imageWidth / 2 - signatureTextLength / 2;
 		enablePlusButtons = isOwner && signature != MAX_SIGNATURE;
 		enableMinusButtons = isOwner && signature != 0;
 		minusThree.active = enableMinusButtons;
@@ -350,6 +363,13 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 		plusOne.active = enablePlusButtons;
 		plusTwo.active = enablePlusButtons;
 		plusThree.active = enablePlusButtons;
+
+		if (!throughTextField) {
+			String textFieldValue = leftPaddedSignature();
+
+			if (!signatureTextField.getValue().equals(textFieldValue))
+				signatureTextField.setValue(textFieldValue);
+		}
 	}
 
 	public void changeLevelState(int i, boolean active) {
@@ -366,5 +386,9 @@ public class KeycardReaderScreen extends AbstractContainerScreen<KeycardReaderMe
 			return Optional.empty();
 		else
 			return Optional.of(message);
+	}
+
+	private String leftPaddedSignature() {
+		return StringUtils.leftPad("" + signature, 5, "0");
 	}
 }