+++ Changelog.md
@@ -8,6 +8,7 @@
 - Change: The model and texture of the Motion-Activated Light have been updated
 - Change: The Reinforced Chain has been renamed to Reinforced Iron Chain to be consistent with vanilla
 - Change: The Trophy System now shows its targeting laser even when disguised
+- Change: The signature of a Keycard Reader can now be changed through a text box
 - API: New IOwnable#isOwnedBy overload taking a Player and a boolean for checking ownership respecting the new Incognito Mask
 - API: New method Owner#isDefaultOwner to check if the Owner object has no player data associated with it
 - API: New method IPasscodeConvertible#getRequiredKeyPanels to control how many key panel items are consumed when converting this block